Seperate Tests into different Classes
