+++ tests/dotnet-structuring.tests/ExecuteTest.cs
@@ -12,6 +12,7 @@ namespace dotnet_structuring.tests
     {
         //Setup Variables
         Variables _ = new Variables();
+        TempDirectory temp = new TempDirectory();
         string CurrentLog;
 
         //Setup EventHandler
@@ -32,7 +33,7 @@ namespace dotnet_structuring.tests
         public void Setup()
         {
             _.ProjectName = "TestProject";
-            _.Directory = new TempDirectory();
+            _.Directory = temp;
             _.Artifacts = "artifacts";
             _.Build = "build";
             _.Docs = "docs";
@@ -70,40 +71,8 @@ namespace dotnet_structuring.tests
             WireEventHandlers(OutputLogs);
             OutputLogs.CreateScript(_.Directories, _.NETCommand, _.ProjectName);
             Assert.AreEqual("Done.", CurrentLog);
+
+            Directory.Delete(temp, true);
         }
-    }
-    class TempDirectory : IDisposable
-    {
-        public TempDirectory()
-        {
-            path = Path.Combine(
-                Path.GetTempPath(),
-                Guid.NewGuid().ToString()
-            );
-            Directory.CreateDirectory(path);
-        }
-
-        readonly string path;
-
-        /// 
-
-        /// Allows the TempDirectory to be used anywhere a string is required.
-        /// 
-        public static implicit operator string(TempDirectory directory)
-        {
-            return directory.path;
-        }
-
-        public override string ToString()
-        {
-            return path;
-        }
-
-        public void Dispose()
-        {
-            Directory.Delete(path, true);
-        }
-
     }
 }
-
